Separate HTTP dashboard routes behind the facade
Continue #687 Wave 2 by moving dashboard REST and Prometheus route handlers out of the HTTP transport facade while keeping the existing HTTPTransport methods as compatibility delegates.

Constraint: #687 requires facade-preserving no-behavior-change god-object splits with reviewable route maps.

Rejected: Mix this into the auth/CORS/config split | dashboard routing is an independent review unit and should stack after PR #1183.

Confidence: high

Scope-risk: moderate

Directive: Keep src/transports/http.ts as the public transport facade until all Wave 2 route/body/session slices land.

Tested: npm run build; npm test -- tests/security/dashboard-rest-authz.test.ts tests/transports/http-bearer-auth.test.ts tests/transports/http-streamable.test.ts tests/transports/prometheus.test.ts --runInBand; npm run lint -- --quiet; npm run lint:tier; git diff --check

Not-tested: Full repository Jest suite and live dashboard browser session.
